Organize code into folders corresponding to program features
