+++ projects/VoxelRenderer/src/graphics/GraphicsUtils.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <GL/glew.h>
-#include <GLFW/glfw3.h>
+#include "GL/glew.h"
+#include "GLFW/glfw3.h"
 
 class GraphicsUtils
 {
