+++ projects/VoxelRenderer/src/rendering/VoxelRenderer.cpp
@@ -1,14 +1,14 @@
 #include "VoxelRenderer.h"
-#include "GraphicsUtils.h"
-#include "ShaderManager.h"
-#include "TupleHasher.h"
-#include "VoxelWorld.h"
-#include <GL/glew.h>
-#include <GLFW/glfw3.h>
+#include "../graphics/GraphicsUtils.h"
+#include "../graphics/ShaderManager.h"
+#include "../utility/TupleHasher.h"
+#include "../world/VoxelWorld.h"
+#include "GL/glew.h"
+#include "GLFW/glfw3.h"
+#include "glm/common.hpp"
+#include "glm/gtc/type_ptr.hpp"
+#include "glm/mat3x3.hpp"
 #include <cmath>
-#include <glm/common.hpp>
-#include <glm/gtc/type_ptr.hpp>
-#include <glm/mat3x3.hpp>
 #include <string>
 #include <tuple>
 #include <unordered_map>
