+++ projects/VoxelRenderer/src/utility/BufferedEvent.h
