+++ projects/VoxelRenderer/src/windowing/Window.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "InputManager.h"
-#include <GLFW/glfw3.h>
-#include <glm/common.hpp>
+#include "../input/InputManager.h"
+#include "GLFW/glfw3.h"
+#include "glm/common.hpp"
 #include <memory>
 #include <unordered_map>
